+++ TP_TEMA6_GRUPO3/src/main/java/frgp/utn/edu/ar/dao/ITurno.java
@@ -11,10 +11,10 @@ public interface ITurno {
 	public boolean Exist(int id);
 	public boolean validarTurno(int legajoM, Date fecha, String hora);
 	public Turno ReadOne(int id);
-	public int informeAusencias(int mes);
-	public int informeTotalTurnos(int mes);
-	public int informePacienteAtendidos(int mes);
-	public double informePorcentAsistencias(int mes);
+	public long informeAusencias(int mes);
+	public long informeTotalTurnos(int mes);
+	public long informePacienteAtendidos(int mes);
+	public Double informePorcentAsistencias(int mes);
 	public boolean Update(Turno turno);
 	public boolean Delete(Turno turno) ;
 	public List<Turno> ReadAll() ;
